[edk2-platforms/devel-IntelAtomProcessorE3900] Flag after GPIO init
Need to prevent modifying the SoC PADs after GPIO init in the I2C code
so I added a PCD flag for this purpose. While in the I2C library, I
cleaned it up. This still leaves a brief window from after GPIO init
and when the flag gets set for the I2C library to modify the SoC I2C
PADs.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Kelly Steele <kelly.steele@intel.com>
Reviewed-by: David Wei <david.wei@intel.com>